Vite Scaffolding
I have fixed a problem regarding the display of this project on other devices.
