+++ vite-project/.gitignore
