+++ vite-project/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "vite-api-project",
+  "name": "vite-project",
   "private": true,
   "version": "0.0.0",
   "type": "module",
